move video-file component into folder group
